requirements.txt
alchemy==20.5
asgiref==3.5.2
attrs==22.1.0
autopep8==1.6.0
beautifulsoup4==4.11.1
black==22.8.0
bs4==0.0.1
certifi==2022.9.14
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
docutils==0.18.1
drf-yasg==1.21.3
filelock==3.0.12
greenlet==1.1.2
idna==3.4
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
lockfile==0.12.2
MarkupSafe==2.1.1
# Editable install with no version control (MyCart==1.0)
-e d:\rain\shoppingcart\gitbas
mypy-extensions==0.4.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
numpy==1.22.3
packaging==21.3
pandas==1.4.2
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyfiglet==0.8.post1
PyJWT==2.4.0
pyparsing==3.0.9
pyperclip==1.8.2
pytest==7.1.3
python-daemon==2.2.4
python-dateutil==2.8.2
pytz==2022.2.1
regex==2022.6.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
